v1.7(999.22a): mechanical-fix warning sweep — 23 sites (Qt6 + sprintf + C4267)
Wave 2 mechanical execution of the 23-site preseed in
.planning/v1.7-WARNING-CENSUS-PRESEED.md. Each group is a one- to two-line
edit with no architectural change; companion phases 999.22c (per-site
review) and 999.29 (GENETICDOCK delete) run in parallel.

Group 1 — sprintf → snprintf (5 sites)
  Adds bounds to the v1.6.2 sweep (commit 85d191d) that missed these.
  - source/COMMON/logStream.C:183                 (local buf[64])
  - source/FORMAT/GAMESSDatFile.C:91              (local line[255])
  - source/FORMAT/KCFFile.C:77, :102              (BALL_MAX_LINE_LENGTH)
  - source/FORMAT/PDBFileDetails.C:1672           (local PDB_date_format[10])

Group 2 — Qt 6 QMessageBox ctor/static deprecation (7 sites)
  Converts 6-arg legacy ctor `QMessageBox(title, text, icon, b0, b1, b2)`
  to Qt 6 `QMessageBox(icon, title, text, StandardButtons)`, and the
  int-button question/critical static overloads to the StandardButtons
  form. Button semantics preserved (single-Ok dialogs → Ok; Yes/No
  question → Yes|No with sensible default).
  - source/VIEW/DIALOGS/dockDialog.C:732, :744
  - source/VIEW/DIALOGS/dockResultDialog.C:397, :449
  - source/VIEW/KERNEL/mainControl.C:2138         (question)
  - source/VIEW/WIDGETS/dockingController.C:110   (question; fix latent
        bug where the bare `if(QMessageBox::question(...))` truth-tested
        the returned StandardButton enum instead of comparing — now
        explicitly compares against QMessageBox::No)
  - source/APPLICATIONS/BALLVIEW/demoTutorialDialog.C:586  (critical)

Group 3 — Qt 6 QMouseEvent::x()/y() local-position deprecation (6 sites)
  evt->x() / evt->y() → static_cast<int>(evt->position().x()) /
  static_cast<int>(evt->position().y()). position() returns QPointF;
  downstream consumers (Vector2 ints, insert_(int,int,...),
  mapViewportTo3D(int,int)) need int.
  - source/VIEW/KERNEL/MODES/editMode.C:279, :280, :364, :456

Group 4 — Qt 6 QMouseEvent ctor deprecation (1 header, ×2 instantiations)
  Converts deprecated 5-arg QMouseEvent ctor to the Qt 6 recommended
  6-arg form with explicit globalPos (using localPos value since the
  TestFramework's synthetic events have local==global by construction).
  - include/BALL/VIEW/WIDGETS/testFramework.h:71

Group 5 — Qt 6 QImage::mirrored deprecation (1 site)
  .mirrored() → .flipped(Qt::Vertical). Default mirrored() = vertical
  flip; preserves the GL byte-order flip for glDrawPixels.
  - source/VIEW/RENDERING/RENDERERS/glRenderer.C:1160

Group 6 — Windows MSVC C4267 size_t→unsigned int narrowing (4 sites)
  Sibling-cluster to v1.6.2 commit 0a75ede (which fixed line 1623 in
  the same function). Vertex (unsigned __int64 on LLP64) → unsigned int
  map key with explicit static_cast at each insert / find / operator[]
  site in the connectedComponents inner loop.
  - source/STRUCTURE/binaryFingerprintMethods.C:1602, :1604, :1608, :1609

Files touched: 13. Local cmake build (BALL VIEW BALLView) on
build/uat-coldstart is green with zero warnings on touched files.

Expected warning-census delta (macOS):
  -5  -Wdeprecated-declarations [sprintf]
  -7  -Wdeprecated-declarations [QMessageBox]
  -6  -Wdeprecated-declarations [QMouseEvent x()/y()]
  -2  -Wdeprecated-declarations [QMouseEvent ctor, 2 TU instantiations]
  -1  -Wdeprecated-declarations [QImage::mirrored]
  ───
  -21 macOS warnings retired
Expected Windows delta: -4 C4267 in binaryFingerprintMethods.C.

+++ source/STRUCTURE/binaryFingerprintMethods.C
@@ -1599,14 +1599,19 @@ bool BinaryFingerprintMethods::connectedComponents(const vector<unsigned int>& s
 		boost::unordered_map<unsigned int, boost::unordered_map<unsigned int, unsigned int> >::iterator ccs_iter;
 		BOOST_FOREACH(Vertex current_vertex, vertices(sim_graph))
 		{
-			if (ccs_tmp.find(ds->find_set(current_vertex)) == ccs_tmp.end())
+			// Explicit narrowing cast from Vertex (boost::graph_traits
+			// vertex_descriptor, unsigned __int64 on Windows LLP64) to
+			// unsigned int — the map key type. Vertex counts never approach
+			// UINT_MAX in practice; sibling-cluster fix to v1.6.2 commit
+			// 0a75ede which fixed the cc_sizes.insert site below.
+			if (ccs_tmp.find(static_cast<unsigned int>(ds->find_set(current_vertex))) == ccs_tmp.end())
 			{
-				ccs_tmp[ds->find_set(current_vertex)] = boost::unordered_map<unsigned int, unsigned int>();
+				ccs_tmp[static_cast<unsigned int>(ds->find_set(current_vertex))] = boost::unordered_map<unsigned int, unsigned int>();
 			}
-			
+
 // 			ccs_tmp[ds->find_set(current_vertex)][current_vertex] = ccs_tmp[ds->find_set(current_vertex)].size();
-			ccs_tmp[ds->find_set(current_vertex)].insert(make_pair(current_vertex, 0));
-			ccs_tmp[ds->find_set(current_vertex)][current_vertex] = ccs_tmp[ds->find_set(current_vertex)].size();
+			ccs_tmp[static_cast<unsigned int>(ds->find_set(current_vertex))].insert(make_pair(static_cast<unsigned int>(current_vertex), 0u));
+			ccs_tmp[static_cast<unsigned int>(ds->find_set(current_vertex))][static_cast<unsigned int>(current_vertex)] = static_cast<unsigned int>(ccs_tmp[static_cast<unsigned int>(ds->find_set(current_vertex))].size());
 		}
